KAFKA-19800: moved PartitionMetadataClient to org.apache.kafka.coordinator.group
